updated auth.rb, class method access_token
